+++ aes/src/ni.rs
@@ -37,7 +37,7 @@ use crate::{Block, Block8};
 use cipher::{
     consts::{U16, U24, U32, U8},
     inout::InOut,
-    AlgorithmName, BlockBackend, BlockCipher, BlockClosure, BlockDecrypt, BlockEncrypt,
+    AlgorithmName, BlockBackend, BlockCipher, BlockCipherDecrypt, BlockCipherEncrypt, BlockClosure,
     BlockSizeUser, Key, KeyInit, KeySizeUser, ParBlocksSizeUser,
 };
 use core::fmt;
@@ -109,13 +109,13 @@ macro_rules! define_aes_impl {
             type BlockSize = U16;
         }
 
-        impl BlockEncrypt for $name {
+        impl BlockCipherEncrypt for $name {
             fn encrypt_with_backend(&self, f: impl BlockClosure<BlockSize = U16>) {
                 self.encrypt.encrypt_with_backend(f)
             }
         }
 
-        impl BlockDecrypt for $name {
+        impl BlockCipherDecrypt for $name {
             fn decrypt_with_backend(&self, f: impl BlockClosure<BlockSize = U16>) {
                 self.decrypt.decrypt_with_backend(f)
             }
@@ -157,6 +157,7 @@ macro_rules! define_aes_impl {
         }
 
         impl KeyInit for $name_enc {
+            #[inline]
             fn new(key: &Key<Self>) -> Self {
                 // SAFETY: we enforce that this code is called only when
                 // target features required by `expand` were properly checked.
@@ -170,7 +171,7 @@ macro_rules! define_aes_impl {
             type BlockSize = U16;
         }
 
-        impl BlockEncrypt for $name_enc {
+        impl BlockCipherEncrypt for $name_enc {
             fn encrypt_with_backend(&self, f: impl BlockClosure<BlockSize = U16>) {
                 f.call(&mut self.get_enc_backend())
             }
@@ -220,6 +221,7 @@ macro_rules! define_aes_impl {
         }
 
         impl KeyInit for $name_dec {
+            #[inline]
             fn new(key: &Key<Self>) -> Self {
                 $name_enc::new(key).into()
             }
@@ -244,7 +246,7 @@ macro_rules! define_aes_impl {
             type BlockSize = U16;
         }
 
-        impl BlockDecrypt for $name_dec {
+        impl BlockCipherDecrypt for $name_dec {
             fn decrypt_with_backend(&self, f: impl BlockClosure<BlockSize = U16>) {
                 f.call(&mut self.get_dec_backend());
             }