+++ threefish/tests/mod.rs
@@ -1,7 +1,7 @@
 //! Test vectors from:
 //! https://github.com/weidai11/cryptopp/blob/master/TestVectors/threefish.txt
-#![cfg(featue = "cipher")]
-use cipher::{Block, BlockDecrypt, BlockEncrypt, KeyInit};
+#![cfg(feature = "cipher")]
+use cipher::{Block, BlockCipherDecrypt, BlockCipherEncrypt, KeyInit};
 use hex_literal::hex;
 use threefish::{Threefish1024, Threefish256, Threefish512};
 
@@ -15,6 +15,7 @@ struct Vector {
 macro_rules! impl_test {
     {$name:ident, $cipher:ty, $tests:expr,} => {
         #[test]
+        #[allow(deprecated)] // uses `clone_from_slice`
         fn $name() {
             let vectors = $tests;
             for &Vector { key, tweak, pt, ct } in vectors.iter() {