+++ magma/CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Bump `cipher` to v0.5.0-pre.1; MSRV 1.65 ([#394])
+- Bump `cipher` dependency to v0.5.0-pre.2 ([#398])
+- Use `BlockCipherEncrypt`/`BlockCipherDecrypt` trait names ([#400])
+- bump `cipher` dependency to `0.5.0-pre.4` ([#413])
+
+[#394]: https://github.com/RustCrypto/block-ciphers/pull/394
+[#398]: https://github.com/RustCrypto/block-ciphers/pull/398
+[#400]: https://github.com/RustCrypto/block-ciphers/pull/400
+[#413]: https://github.com/RustCrypto/block-ciphers/pull/413
+
+## 0.9.0 (2023-08-06)
+### Breaking changes
+- API of the `Sbox` trait is changed, S-Box expansion is now performed
+internally, helper methods are removed. Most users of the crate should not be
+affected by this change. ([#376])
+
+[#376]: https://github.com/RustCrypto/block-ciphers/pull/376
+
 ## 0.8.1 (2022-02-17)
 ### Fixed
 - Minimal versions build ([#303])